+++ src/clirun.nim
@@ -1,10 +1,11 @@
 from std/os import fileExists, commandLineParams, parentDir
-from std/json import parseJson, `{}`, getStr, items
+from std/json import parseJson, `{}`, getStr, items, keys
+from std/tables import Table, `[]`, `[]=`, `$`, pairs
 
 from std/strutils import join
 from std/osproc import execCmd
 
-from clirun/utils import evalPath
+from clirun/utils import parseVars
 
 const debugging = false
 
@@ -16,19 +17,22 @@ else:
 
 type
   Config = object
-    secret, fileData, destination: string
+    secret: string
+    files: Table[string, string]
     commands: seq[string]
 
 proc getConfig(json: string): Config =
-  let
-    node = parseJson json
-    filePath = node{"file"}.getStr
+  let node = parseJson json
 
   result = Config(
     secret: encode node{"secret"}.getStr,
-    fileData: encode readFile filePath,
-    destination: encode node{"destination"}.getStr,
   )
+  for file in node{"files"}.keys:
+    let dest = encode node{"files", file}.getStr
+    echo "Encoding " & file
+    result.files[dest] = encode readFile file
+    echo "Successfully encoded " & $result.files[dest].len & " bytes" # TODO THIS and copu theses files and add a option to disable encoding
+
   for command in node{"commands"}:
     result.commands.add encode command.getStr
 
@@ -37,16 +41,14 @@ when conf.len == 0:
   {.fatal: "Please provide config path with `-d:conf=/path/to/conf.json`".} 
 
 when isMainModule:
-  static: echo "Encoding the binary"
   const config = getConfig staticRead conf
-  static: echo "Successfully encoded " & $config.fileData.len & " bytes" # TODO THIS and copu theses files and add a option to disable encoding
 
   let param = commandLineParams().join " "
 
   if param == config.secret.decode:
-    var filename = evalPath decode config.destination
-    writeFile filename, decode config.fileData
+    for dest, data in config.files:
+      dest.decode.parseVars.writeFile decode data
     for cmd in config.commands:
-      let code = execCmd(decode cmd)
+      let code = execCmd(parseVars decode cmd)
       if code != 0:
         quit code