+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 0.3.0 (Aug 9, 2022)
+
+- Added env vars parsing to commands
+- Added multiple files encoding
+
+---
+
 ## Version 0.2.1 (2022/06/21)
 
 - Fixed getting config path in compilation